requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Web Interface
gradio>=4.0.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bar
tqdm>=4.65.0

# Metrics
scikit-learn>=1.3.0

# Data Processing
pandas>=2.0.0
